+++ Assets/Scripts/Boss/Damageable.cs
@@ -3,8 +3,8 @@ using System.Collections;
 
 public class Damageable : MonoBehaviour {
 
-    public int maxHealth;
-    public int currentHealth;
+    public float maxHealth;
+    public float currentHealth;
 	public int bodyDamage;
 	public string damagedBy; //PlayerProjectile
     public bool dead = false;
@@ -20,7 +20,7 @@ public class Damageable : MonoBehaviour {
 		}
 	}
 
-	public virtual void takeDamage(int damage){
+	public virtual void takeDamage(float damage){
         currentHealth -= damage;
         if (currentHealth <= 0) {
             Destroy(gameObject);
